% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.R
\name{alpha_q10_mutsel_balance}
\alias{alpha_q10_mutsel_balance}
\title{Get the alpha and q10 values}
\usage{
alpha_q10_mutsel_balance(pa, mu, gamma, omega, tol = 0.001)
}
\arguments{
\item{pa}{Proportion of ancestral in the population}

\item{mu}{Transition from Variant to Ancestral or molecular clock rate}

\item{gamma}{Recovery rate}

\item{omega}{Relative transmission fitness}

\item{tol}{Tolerance}
}
\value{
The valeu of alpha = q10/q01. q10 is the transition from ancestral to variant
and q01 is the transition from variant to ancestral or molecular clock rate.
It represents the inflation of the mutation rate from A to V.
}
\description{
Get the alpha and q10 values
}
\examples{
alpha_q10_mutsel_balance( pa = 0.85, mu = 0.0016, gamma = 1/10.2, omega = 0.9, tol = 1e-3 )
}
