% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{get_CI}
\alias{get_CI}
\title{Get confidence interval}
\usage{
get_CI(x, MAXCOEFVAR = 1.5)
}
\arguments{
\item{x}{Object of class bissecofit.}

\item{MAXCOEFVAR}{Threshold value for approximate CIs.}
}
\value{
dataframe with the estimated values and their confidence interval
}
\description{
Get confidence interval
}
